+++ src/ui/instructionsList.ts
@@ -1,69 +1,48 @@
-import type {
-    Webview, 
-    WebviewViewProvider, 
-    WebviewView, 
-    WebviewViewResolveContext, 
-    CancellationToken,
-    ExtensionContext
-} from 'vscode'
+import type { Webview, WebviewViewProvider, WebviewView,ExtensionContext } from 'vscode'
+import type { InstructionList, Message } from '@/types'
+import type { Parser } from '@/parser'
 
-import { window, Uri, workspace } from 'vscode'
+import { window, commands } from 'vscode'
 import { getUri, getNonce } from '@/util'
-import { ASMInstruction, Message } from '@/types'
+import { openInstructionViewCommand } from '@/commands/openInstructionView'
 
 
 export const instructionsListWebviewId: string = `instructionsList`
 
-export const initInstructionsListWebview = (context: ExtensionContext) => {
-    const webviewProvider = new InstructionsListWebview(context)
+export const initInstructionsListWebview = (
+    context: ExtensionContext,
+    parser: Parser
+) => {
+    const webviewProvider = new InstructionsListWebview(context, parser)
     window.registerWebviewViewProvider(instructionsListWebviewId, webviewProvider)
     return webviewProvider
 }
 
 export class InstructionsListWebview implements WebviewViewProvider {
 
-    private data: {
-        [key: string]: ASMInstruction[]
-    } | null = null
-
-    private readData() {
-        workspace.fs.readFile(
-            Uri.joinPath(this.context.globalStorageUri, 'instructionsList.json')
-        ).then(
-            (array) => {
-                this.data = JSON.parse(Buffer.from(array).toString('utf-8')) as {
-                    [name: string]: ASMInstruction[]
-                }
-                this.sendDataMessage()
-            }, 
-            (error) => {
-                console.log(`InstructionsListWebview can't read data:\n${error}`)
-            }
-        )
-    }
-
-    refresh() {
-        this.readData()
-    }
+    private data: InstructionList = null
 
     private view: WebviewView | null = null
 
     constructor (
-        private context: ExtensionContext
+        private context: ExtensionContext,
+        private parser: Parser
     ) {
         
     }
 
+    public async refresh() {
+        this.data = await this.parser.getInstructionsList()
+        this.sendDataMessage()
+    }
+
     public async resolveWebviewView (
-        webviewView: WebviewView,
-        context: WebviewViewResolveContext,
-        token: CancellationToken
+        webviewView: WebviewView
     ) {
         this.view = webviewView
         
 		webviewView.webview.options = {
-			enableScripts: true,
-            localResourceRoots: [getUri(webviewView.webview, this.context.extensionUri, ['dist'])]
+			enableScripts: true
 		}
 
 		webviewView.webview.html = await this.getWebviewContent(webviewView.webview)
@@ -73,11 +52,11 @@ export class InstructionsListWebview implements WebviewViewProvider {
     private async getWebviewContent(
         webview: Webview
     ) {
-        const scriptUri = getUri(webview, this.context.extensionUri, ['webview.js'])
-        const vscodeUIUrl = getUri(webview, this.context.extensionUri, ['vscodeUI.js'])
+        const scriptUri = getUri(webview, this.context.extensionUri, ['dist', 'webview.js'])
+        const vscodeUIUrl = getUri(webview, this.context.extensionUri, ['dist', 'vscodeUI.js'])
 
-        const codiconsUri = getUri(webview, this.context.extensionUri, ['codeicon.css'])
-        const styleUri = getUri(webview, this.context.extensionUri, ['instructionsList.css'])
+        const codiconsUri = getUri(webview, this.context.extensionUri, ['dist', 'codicon.css'])
+        const styleUri = getUri(webview, this.context.extensionUri, ['media', 'instructionsList.css'])
         
         const nonce = getNonce()
 
@@ -132,9 +111,11 @@ export class InstructionsListWebview implements WebviewViewProvider {
 
                 if (type == 'loaded')
                     this.sendDataMessage()
+                else if (type == 'selected')
+                    commands.executeCommand(openInstructionViewCommand, message.data)
+                
             }
         )
-
     }
 
     private sendDataMessage () {
@@ -143,4 +124,5 @@ export class InstructionsListWebview implements WebviewViewProvider {
             data: this.data
         } as Message)
     }
+
 }